secret-api/requirements.txt
flask==2.3.3
Jinja2>=3.1.2
psycopg2==2.9.6
itsdangerous==2.1.2
Werkzeug==2.3.7
Markupsafe==2.1.3
Click==8.1.3
SQLAlchemy==2.0.21
flask-marshmallow==0.14.0
flask_sqlalchemy
marshmallow-sqlalchemy==0.24.1
typing-extensions==4.8.0
greenlet==2.0.2

pip~=23.2.1
cryptography~=41.0.1
blinker~=1.6.2
pycparser~=2.21
setuptools~=68.0.0
zipp~=3.15.0
python-dotenv~=1.0.0